requirements.txt
flask==3.1.0
flask-cors==5.0.1
google-api-python-client==2.163.0
python-dotenv==1.1.0
